Revert "Removed some random matlab generated files. Removed some of the now unnecessary flight modules. Converted everything to f64."
This reverts commit b4c5a636d59450568379fb3aa121f8b5c0b05f17.

+++ copter/src/hardware/mock.rs
@@ -54,7 +54,7 @@ impl Magnetometer for MockSensor {
 impl MotorManager for MockSensor {
     fn arm(&mut self) {}
     fn terminate(&mut self) {}
-    fn set_powers(&mut self, powers: [f64; 4]) {}
+    fn set_powers(&mut self, powers: [f32; 4]) {}
     fn process_command(&mut self, command: MotorCommand) {}
     fn calibrate(&mut self) {}
 }
